feat: modularize code, making __init__.py smaller in size
