+++ files/ReviewerBottomBar.css
@@ -1,5 +1,6 @@
 html {
   height: 100%;
+  min-height: 65px;
 }
 body {
   overflow: hidden;
@@ -37,7 +38,7 @@ td[align="center"] > span {
   position: absolute;
   display: block;
   left: 50%;
-  transform: translate(-50%, -13px);
+  transform: translate(-50%, -10px);
   color: var(--text-fg, var(--legacy-text-fg));
 }
 body .stattxt#time {
